app/requirements.txt
django-environ<=0.12.0
djangorestframework<=3.16.1
djangorestframework-xml<=2.0.0
pysam<=0.23.3
pytest-django<=4.11.1
